refactor: make constructors for main with classess
- for visablility and scalability
